+++ libtiff/tif_next.c
@@ -127,8 +127,8 @@ NeXTDecode(TIFF* tif, uint8_t* buf, tmsize_t occ, uint16_t s)
 				if (npixels >= imagewidth)
 					break;
                 if (op_offset >= scanline ) {
-                    TIFFErrorExt(tif->tif_clientdata, module, "Invalid data for scanline %ld",
-                        (long) tif->tif_row);
+                    TIFFErrorExt(tif->tif_clientdata, module, "Invalid data for scanline %"PRIu32,
+                        tif->tif_row);
                     return (0);
                 }
 				if (cc == 0)
@@ -144,8 +144,8 @@ NeXTDecode(TIFF* tif, uint8_t* buf, tmsize_t occ, uint16_t s)
 	tif->tif_rawcc = cc;
 	return (1);
 bad:
-	TIFFErrorExt(tif->tif_clientdata, module, "Not enough data for scanline %ld",
-	    (long) tif->tif_row);
+	TIFFErrorExt(tif->tif_clientdata, module, "Not enough data for scanline %"PRIu32,
+	    tif->tif_row);
 	return (0);
 }
 
@@ -158,7 +158,7 @@ NeXTPreDecode(TIFF* tif, uint16_t s)
 
 	if( td->td_bitspersample != 2 )
 	{
-		TIFFErrorExt(tif->tif_clientdata, module, "Unsupported BitsPerSample = %d",
+		TIFFErrorExt(tif->tif_clientdata, module, "Unsupported BitsPerSample = %"PRIu16,
 					 td->td_bitspersample);
 		return (0);
 	}
